Refactor - Enables concurrent strategies and remote start/stop.

+++ strategies/AbstractStrategy.py
@@ -43,7 +43,7 @@ class AbstractStrategy(ABC):
         message = f"[{now}][{action}]: {token_amount} ETH at ${price}"
         failure_message = f"[{now}][FAILURE]: Unable to {action} {token_amount} ETH at ${price}"
 
-        transaction_value = round(price * token_amount, 2)
+        transaction_value = float(round(price * token_amount, 2))
         token_amount = round(token_amount, 5)
 
         if action == Signal.BUY:
